# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Divine Hindu Inventory Tracking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory balances with product and warehouse info
      parameters:
        - in: query
          name: warehouseId
          schema: { type: integer }
        - in: query
          name: productId
          schema: { type: integer }
        - in: query
          name: lowStock
          schema: { type: boolean }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Inventory balances
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, limit]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/InventoryBalance" }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }

  # ── Stock Movements ───────────────────────────────────────
  /inventory/movements:
    get:
      operationId: listMovements
      tags: [inventory]
      summary: List stock movements
      parameters:
        - in: query
          name: productId
          schema: { type: integer }
        - in: query
          name: warehouseId
          schema: { type: integer }
        - in: query
          name: movementType
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, limit]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/StockMovement" }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
    post:
      operationId: createMovement
      tags: [inventory]
      summary: Create a stock movement (inward, sale, transfer, damage, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MovementInput" }
      responses:
        "201":
          description: Movement created and ledger updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockMovement" }
        "409":
          description: Duplicate movement ID (idempotency)
        "422":
          description: Insufficient stock

  # ── Inventory Ledger ──────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List sales orders
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: channel
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, limit]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/SalesOrder" }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a sales order (triggers stock reservation)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrder" }
        "422":
          description: Insufficient stock
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID with items
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrder" }

  # ── Returns ───────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, skuCode, productName, category, brand, unit, minimumStock, reorderStock, isActive, createdAt]
      properties:
        id: { type: integer }
        skuCode: { type: string }
        productName: { type: string }
        category: { type: string }
        brand: { type: string }
        unit: { type: string }
        minimumStock: { type: integer }
        reorderStock: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: ["string", "null"], format: date-time }
...
    Warehouse:
      type: object
      required: [id, warehouseName, location, warehouseType, createdAt]
      properties:
        id: { type: integer }
        warehouseName: { type: string }
        location: { type: string }
        warehouseType: { type: string, enum: [MAIN, QUARANTINE, DAMAGED] }
        createdAt: { type: string, format: date-time }
...
    InventoryBalance:
      type: object
      required: [id, productId, warehouseId, availableQty, reservedQty, damagedQty, expiredQty, returnedQty, lastUpdated]
      properties:
        id: { type: integer }
        productId: { type: integer }
        warehouseId: { type: integer }
        availableQty: { type: integer }
        reservedQty: { type: integer }
        damagedQty: { type: integer }
        expiredQty: { type: integer }
        returnedQty: { type: integer }
        lastUpdated: { type: string, format: date-time }
        product: { $ref: "#/components/schemas/Product" }
        warehouse: { $ref: "#/components/schemas/Warehouse" }
    StockMovement:
      type: object
      required: [id, movementId, productId, warehouseId, movementType, qty, referenceType, referenceId, createdAt]
      properties:
        id: { type: integer }
        movementId: { type: string }
        productId: { type: integer }
        warehouseId: { type: integer }
        movementType:
          type: string
          enum: [INWARD, SALE, RETURN, DISPATCH, TRANSFER_IN, TRANSFER_OUT, DAMAGE, EXPIRED, ADJUSTMENT]
        qty: { type: integer }
        referenceType: { type: string }
        referenceId: { type: string }
        sourceLocation: { type: ["string", "null"] }
        destinationLocation: { type: ["string", "null"] }
        createdBy: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        product: { $ref: "#/components/schemas/Product" }
        warehouse: { $ref: "#/components/schemas/Warehouse" }
    MovementInput:
      type: object
      required: [movementId, productId, warehouseId, movementType, qty, referenceType, referenceId]
      properties:
        movementId: { type: string }
        productId: { type: integer }
        warehouseId: { type: integer }
        movementType:
          type: string
          enum: [INWARD, SALE, RETURN, DISPATCH, TRANSFER_IN, TRANSFER_OUT, DAMAGE, EXPIRED, ADJUSTMENT]
        qty: { type: integer }
        referenceType: { type: string }
        referenceId: { type: string }
        sourceLocation: { type: string }
        destinationLocation: { type: string }
        createdBy: { type: string }

    # ── Ledger schemas ────────────────────────────────────────
...
    SalesOrderItem:
      type: object
      required: [id, orderId, productId, warehouseId, qty]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        productId: { type: integer }
        warehouseId: { type: integer }
        qty: { type: integer }
        product: { $ref: "#/components/schemas/Product" }
    SalesOrder:
      type: object
      required: [id, orderNo, channel, status, createdAt]
      properties:
        id: { type: integer }
        orderNo: { type: string }
        channel: { type: string, enum: [WEBSITE, AMAZON, FLIPKART, POS, OTHER] }
        status: { type: string, enum: [PENDING, CONFIRMED, DISPATCHED, CANCELLED] }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/SalesOrderItem" }
    OrderItemInput:
      type: object
      required: [productId, warehouseId, qty]
      properties:
        productId: { type: integer }
        warehouseId: { type: integer }
        qty: { type: integer }
    OrderInput:
      type: object
      required: [channel, items]
      properties:
        channel: { type: string, enum: [WEBSITE, AMAZON, FLIPKART, POS, OTHER] }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItemInput" }

    # ── Return schemas ────────────────────────────────────────
...