# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Divine Hindu Inventory Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  skuCode: string;
  productName: string;
  category: string;
  brand: string;
  unit: string;
  minimumStock: number;
  reorderStock: number;
  isActive: boolean;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
}