# lib/api-zod/src/generated/types/salesOrder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Divine Hindu Inventory Tracking System API
 * OpenAPI spec version: 0.1.0
 */
import type { SalesOrderChannel } from './salesOrderChannel';
import type { SalesOrderItem } from './salesOrderItem';
import type { SalesOrderStatus } from './salesOrderStatus';

export interface SalesOrder {
  id: number;
  orderNo: string;
  channel: SalesOrderChannel;
  status: SalesOrderStatus;
  createdAt: Date;
  items?: SalesOrderItem[];
}