# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Divine Hindu Inventory Tracking System API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';
import type { StockMovementMovementType } from './stockMovementMovementType';
import type { Warehouse } from './warehouse';

export interface StockMovement {
  id: number;
  movementId: string;
  productId: number;
  warehouseId: number;
  movementType: StockMovementMovementType;
  qty: number;
  referenceType: string;
  referenceId: string;
  /** @nullable */
  sourceLocation?: string | null;
  /** @nullable */
  destinationLocation?: string | null;
  /** @nullable */
  createdBy?: string | null;
  createdAt: Date;
  product?: Product;
  warehouse?: Warehouse;
}